Renamed parser to be the same as the repo name.
